decode: Synchronize reorg's work with cuvid's.
cuvid does its work in the default stream.  So that we can enqueue
the copy whilst the kernel runs, we use a separate and
unsynchronized stream to reorganize the NV12 data into RGB.  With
this commit, we may still potentially enqueue the kernel more
quickly than cuvid can finish, but the event will guarantee that
the kernel will not actually start until cuvid has finished.

This is needed because cuvid internally seems to use multiple
buffers: if it hasn't finished by the time we request the internal
buffer pointer, then it may feed us stale data (from the previous
frame, in our case).

Thanks to Olivier Lapicque for debugging / diagnosing this one.