New API: nvpipe_ptx_path for runtime paths.
We need to load PTX files at runtime for the conversion to/from
NV12 (i.e. YUV).  In most situations, the PREFIX setup during
install time is communicated via config.nvp.h and this suffices.
However if binary is relocated this path may not be invalid.  Some
applications need this API to be able to find shaders in that kind
of environment.