Make it easy to use the Video Codec SDK.  Or not.
This adds a CMake option to let the user select an SDK they wish
to compile against.  Using later SDKs is preferred because they
support additional features that we can take advantage of.
However, they also come with more stringent driver version
requirements.

The previous patch added the header from the 5.0.1 release of the
SDK.  This will be used as the default if the user does not supply
a path.