decode: support output to device buffers.
This checks whether the output pointer given to us during decode
is a device or host pointer.  If it's a device pointer, we avoid
the final copy and simply do the final nv12-to-RGB reorganization
directly into the user's buffer.

This also reorganizes the reorganization code.  A new method
handles where to do the reorganization into and whether that is a
host or device pointer.