After discussion with others, it was deemed that we should not support input from device memory.  The issue is that currently the video codec SDK APIs *only* support input from host memory.  As such, we needed to do a synchronous copy before starting our work.
While convenient for the user, it was deemed better to force them
to do the copy themselves, in the hope that they will do so
intelligently and asynchronously, with a sync point before calling
NVPipe.

We should revisit if/when the video codec SDK ever supports device
memory for input data on decode.  In the meantime, at least input
data is rather small.

Note no change for output data: we can and still do support
efficiently pushing that to a device buffer.